refactor : 프로젝트 폴더 구조 수정
- 폴더 구조를 app 아래 project, recruitment로 구분
