+++ frontend/src/app/layout.tsx
@@ -2,7 +2,6 @@ import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import Provider from '../components/general/Provider'
-import NavBar from '../components/general/NavBar'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -19,9 +18,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <Provider>
-          <NavBar>{children}</NavBar>
-        </Provider>
+        <Provider>{children}</Provider>
       </body>
     </html>
   )
