+++ frontend/src/app/project/(user)/registration/error.tsx
