+++ frontend/src/app/project/layout.tsx
@@ -1,0 +1,9 @@
+import ProjectNavBar from '../../components/general/ProjectNavBar'
+
+export default function ProjectLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return <ProjectNavBar>{children}</ProjectNavBar>
+}
