+++ frontend/src/app/recruitment/layout.tsx
@@ -1,0 +1,9 @@
+import RecruitmentNavBar from '../../components/general/RecruitmentNavBar'
+
+export default function ProjectLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return <RecruitmentNavBar>{children}</RecruitmentNavBar>
+}
