+++ frontend/src/components/general/RecruitmentNavBar.tsx
@@ -1,0 +1,7 @@
+export default function RecruitmentNavBar({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return <div>{children}</div>
+}
